+++ packages/dash-core-components/src/components/Link.react.js
@@ -47,7 +47,7 @@ export default class Link extends Component {
             window.location.pathname = href;
         } else {
             window.history.pushState({}, '', href);
-            window.dispatchEvent(new CustomEvent('onpushstate'));
+            window.dispatchEvent(new CustomEvent('_dashprivate_pushstate'));
         }
         // scroll back to top
         window.scrollTo(0, 0);