+++ packages/dash-core-components/src/components/Location.react.js
@@ -1,6 +1,8 @@
 import {Component} from 'react';
 import PropTypes from 'prop-types';
 import {type} from 'ramda';
+
+import {History} from '@plotly/dash-component-plugins';
 /* global window:true */
 
 /**
@@ -11,6 +13,7 @@ export default class Location extends Component {
     constructor(props) {
         super(props);
         this.updateLocation = this.updateLocation.bind(this);
+        this.onLocationChange = this.onLocationChange.bind(this);
     }
 
     updateLocation(props) {
@@ -81,23 +84,25 @@ export default class Location extends Component {
         }
     }
 
-    componentDidMount() {
-        const listener = () => {
-            return () => {
-                const {setProps} = this.props;
-                setProps({
-                    pathname: window.location.pathname,
-                    href: window.location.href,
-                    hash: window.location.hash,
-                    search: window.location.search,
-                });
-            };
-        };
-        window.addEventListener('onpopstate', listener());
-        window.onpopstate = listener('POP');
+    onLocationChange() {
+        const {setProps} = this.props;
+        setProps({
+            pathname: window.location.pathname,
+            href: window.location.href,
+            hash: window.location.hash,
+            search: window.location.search,
+        });
 
-        // non-standard, emitted by Link.react
-        window.addEventListener('onpushstate', listener());
+        History.dispatchChangeEvent();
+    }
+
+    componentDidMount() {
+        window.onpopstate = this.onLocationChange;
+
+        window.addEventListener(
+            '_dashprivate_pushstate',
+            this.onLocationChange
+        );
         this.updateLocation(this.props);
     }